[#25] SetNicknameViewModel 설계 + 레거시 코드의 메타데이터 수정
